재고와 주문 관련 테이블 재설계 및 후작업중

+++ src/main/java/com/got/service/FileService.java
@@ -4,32 +4,34 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+import javax.inject.Inject;
+
 import org.apache.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.got.dao.FileDao;
 import com.got.enums.MenuLevel;
-import com.got.util.CommonUtil;
+import com.got.mapper.files.FileMapper;
 import com.got.util.FileUtil;
 import com.got.vo.FileVO;
 import com.got.vo.GoodsImageVO;
-import com.got.vo.goods.CategoryVO;
 
 @Service
 public class FileService {
 	
-	@Autowired private FileDao dao;
-	static Logger logger = Logger.getLogger(FileService.class);
+	static Logger log = Logger.getLogger(FileService.class);
+	@Inject private FileMapper fileMapper;
 	
 	public FileVO saveFileInTempPath(String path, MultipartFile uploadFile) {
+		Objects.requireNonNull(uploadFile);
+		if(uploadFile.isEmpty())
+			throw new RuntimeException("업로드된 파일 내용물이 비어있음");
 		FileVO f = FileUtil.saveFileInTempPath(path, uploadFile);
 		try{
-			dao.insert(f);
+			fileMapper.insert(f);
 		}catch(Exception e) {
 			deleteFile(f);
-			logger.error(e);
+			log.error(e);
 			throw new RuntimeException("데이터베이스에 파일정보 저장이 실패하여 앞서 업로드된 파일 삭제 수행");
 		}
 		f.setSave_path(null);
@@ -39,15 +41,7 @@ public class FileService {
 	public List<FileVO> saveFileInTempPath(String path, List<MultipartFile> files) {
 		List<FileVO> fileVOs = new ArrayList<>();
 		files.forEach(file -> {
-			FileVO f = FileUtil.saveFileInTempPath(path, file);
-			try{
-				dao.insert(f);
-			}catch(Exception e) {
-				deleteFile(f);
-				logger.error(e);
-				throw new RuntimeException("데이터베이스에 파일정보 저장이 실패하여 앞서 업로드된 파일 삭제 수행");
-			}
-			f.setSave_path(null);
+			FileVO f = saveFileInTempPath(path, file);
 			fileVOs.add(f);
 		});
 		return fileVOs;
@@ -60,16 +54,6 @@ public class FileService {
 		else
 			throw new RuntimeException("삭제 실패, 확인바람");
 	}
-
-	public List<GoodsImageVO> setupImages(Integer c_no, String[] fileInfoJSON) {
-		Objects.requireNonNull(c_no);
-		if(Objects.isNull(fileInfoJSON))
-			return new ArrayList<>();
-		
-		String folderName = getSaveFolderName(c_no);
-		List<GoodsImageVO> tempGoodsImgs = CommonUtil.getVO(fileInfoJSON, GoodsImageVO.class);
-		return FileUtil.moveToSavePath(folderName, tempGoodsImgs);
-	}
 	
 	public List<GoodsImageVO> setupImages(Integer c_no, List<GoodsImageVO> tempImages) {
 		Objects.requireNonNull(c_no);
@@ -79,7 +63,7 @@ public class FileService {
 		return FileUtil.moveToSavePath(folderName, tempImages);
 	}
 	
-	private String getSaveFolderName(Integer g_no) {
-		return MenuLevel.findBigCategory(g_no).getTitle();
+	private String getSaveFolderName(Integer c_no) {
+		return MenuLevel.findBigCategory(c_no).getTitle();
 	}
 }
